Add tests for IdEmbedder
